+++ src/sg_vpd.c
@@ -677,8 +677,8 @@ decode_dev_ids(const char * print_if_found, unsigned char * buff, int len,
 {
     int m, p_id, c_set, piv, assoc, desig_type, i_len;
     int ci_off, c_id, d_id, naa, vsi, printed, off, u;
-    unsigned long long vsei;
-    unsigned long long id_ext;
+    uint64_t vsei;
+    uint64_t id_ext;
     const unsigned char * ucp;
     const unsigned char * ip;
 
@@ -950,7 +950,7 @@ static void
 decode_transport_id(const char * leadin, unsigned char * ucp, int len)
 {
     int format_code, proto_id, num, j, k;
-    unsigned long long ull;
+    uint64_t ull;
     int bump;
 
     for (k = 0, bump= 24; k < len; k += bump, ucp += bump) {
